+++ Engine/Fleur/Windows/InputWin.h
@@ -13,11 +13,11 @@ protected:
     virtual bool IsKeyPressedImpl(EKeyCode keyCode) const override;
 
     virtual bool IsMouseButtonPressedImpl(EMouseCode mouseCode) override;
-    virtual bool IsMouseWheelScrolledImpl(std::pair<float, float>& pair) const override;
-    virtual std::pair<float, float> GetMousePositionImpl() const override;
-    virtual float GetMouseXImpl() const override;
-    virtual float GetMouseYImpl() const override;
-    virtual glm::vec2 GetMouseDirImpl() const override;
+    virtual bool IsMouseWheelScrolledImpl(std::pair<int, int>& pair) const override;
+    virtual std::pair<int, int> GetMousePositionImpl() const override;
+    virtual int GetMouseXImpl() const override;
+    virtual int GetMouseYImpl() const override;
+    virtual glm::ivec2 GetMouseDirImpl() const override;
 
     Input::KeyInfo m_LastKey;
     Input::MouseInfo m_LastMouse;
